To stop removing a popular STRIPEJS payment module during  upgrade
Hello,

I'm the contributor for the module "Stripejs" payment module and I've got lots of complains that the module does gets deleted while upgrading Prestashop. So I opened an issue to fix it but nothing was done after long time.
https://github.com/PrestaShop/PrestaShop/issues/27561
So I decided to do it myself.

The code which I removed from this autupgrade module, is responsible to delete my module during the upgrade. 

Please consider this change at high priority because thousands of customers are constantly facing this serious issue with  my module.
Best Regards,
Sitender